Factor out code to create Mongoose conn
