+++ src/services/cogrpc/server.spec.ts
@@ -28,7 +28,6 @@ jest.mock('grpc', () => {
 const BASE_ENV_VARS = {
   COGRPC_ADDRESS: 'https://cogrpc.example.com/',
   GATEWAY_KEY_ID: 'base64-encoded key id',
-  MONGO_URI: 'mongo://example.com',
   NATS_CLUSTER_ID: 'nats-cluster-id',
   NATS_SERVER_URL: 'nats://example.com',
   PARCEL_STORE_BUCKET: 'bucket-name',
@@ -38,7 +37,6 @@ const mockEnvVars = configureMockEnvVars(BASE_ENV_VARS);
 describe('runServer', () => {
   test.each([
     'GATEWAY_KEY_ID',
-    'MONGO_URI',
     'NATS_SERVER_URL',
     'NATS_CLUSTER_ID',
     'COGRPC_ADDRESS',
@@ -106,7 +104,6 @@ describe('runServer', () => {
     expect(makeServiceImplementationSpy).toBeCalledWith({
       cogrpcAddress: BASE_ENV_VARS.COGRPC_ADDRESS,
       gatewayKeyIdBase64: BASE_ENV_VARS.GATEWAY_KEY_ID,
-      mongoUri: BASE_ENV_VARS.MONGO_URI,
       natsClusterId: BASE_ENV_VARS.NATS_CLUSTER_ID,
       natsServerUrl: BASE_ENV_VARS.NATS_SERVER_URL,
       parcelStoreBucket: BASE_ENV_VARS.PARCEL_STORE_BUCKET,
